Add properties before corrections
Adding properties before corrections allows the corrections to be the final say
on the records.  This was seen for ligo, which has a VO name ligo, and a project
name LIGO.  Since addProperties adds sets the VO to the ProjectName, the
ProjectName was being set to the lower case ligo.  We have a correction from
lower case to upper case, but the add properties was after the correction stage.